@@ -68,10 +68,11 @@
     <maven.build.timestamp.format>yyyy-MM-dd'T'HH:mm:ssZ</maven.build.timestamp.format>
     <build.time.xsd>${maven.build.timestamp}</build.time.xsd>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <java.version>11</java.version>
+    <java.version>17</java.version>
 
-    <ver.jena>4.10.0</ver.jena>
-    <ver.jetty>10.0.19</ver.jetty>
+    <ver.jena>5.0.0-SNAPSHOT</ver.jena>
+    <ver.jetty>12.0.2</ver.jetty>
+
     <!-- Ensure we get a version compatible with Jena. -->
     <ver.commons-io>2.15.1</ver.commons-io>
     <ver.junit>4.13.2</ver.junit>
