@@ -21,9 +21,9 @@ import java.io.IOException ;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
-import javax.servlet.http.HttpServlet ;
-import javax.servlet.http.HttpServletRequest ;
-import javax.servlet.http.HttpServletResponse ;
+import jakarta.servlet.http.HttpServlet ;
+import jakarta.servlet.http.HttpServletRequest ;
+import jakarta.servlet.http.HttpServletResponse ;
 
 import org.apache.jena.atlas.json.JSON;
 import org.apache.jena.fuseki.server.RequestLog ;
