@@ -19,8 +19,8 @@ package org.seaborne.delta.server.http;
 
 import java.io.IOException;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 import org.seaborne.delta.link.DeltaLink;
 
