@@ -27,8 +27,8 @@ import java.io.PrintStream ;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import javax.servlet.http.HttpServletRequest ;
-import javax.servlet.http.HttpServletResponse ;
+import jakarta.servlet.http.HttpServletRequest ;
+import jakarta.servlet.http.HttpServletResponse ;
 
 import org.apache.jena.atlas.RuntimeIOException;
 import org.apache.jena.atlas.io.IndentedWriter;
