@@ -20,10 +20,10 @@ package org.seaborne.delta.server.http;
 import java.io.IOException ;
 import java.io.File ;
 
-import javax.servlet.ServletOutputStream ;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest ;
-import javax.servlet.http.HttpServletResponse ;
+import jakarta.servlet.ServletOutputStream ;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest ;
+import jakarta.servlet.http.HttpServletResponse ;
 
 import org.apache.jena.riot.WebContent ;
 import org.apache.jena.riot.web.HttpNames ;
