@@ -19,10 +19,10 @@ package org.seaborne.delta.server.http;
 
 import java.io.IOException ;
 
-import javax.servlet.ServletOutputStream ;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest ;
-import javax.servlet.http.HttpServletResponse ;
+import jakarta.servlet.ServletOutputStream ;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest ;
+import jakarta.servlet.http.HttpServletResponse ;
 
 import org.apache.jena.atlas.io.IndentedWriter ;
 import org.apache.jena.atlas.json.JSON ;
