@@ -19,9 +19,9 @@ package org.seaborne.delta.server.http;
 
 import java.io.IOException ;
 
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest ;
-import javax.servlet.http.HttpServletResponse ;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest ;
+import jakarta.servlet.http.HttpServletResponse ;
 
 import org.apache.jena.web.HttpSC ;
 import org.seaborne.delta.Delta ;
